Refactor to separate concerns
Database operations are now handled in db.go, which defines a simple
interface for loading and saving documents.

Document-specific business logic is now defined in document.go, which
also exposes a constructor for creating documents.

The actual service which makes uses of this code lives in pasteburn.go.

HTTP serving functions have also been separated to allow the injection
of a Pasteburn service, as well as passing a context down. This isn't
currently used for anything.

Tests have been refactored to break dependencies on database logic and
I've written a small DB mock for future tests related to the service
itself.